requirements.txt
breeze-connect>=1.0.55
python-telegram-bot>=20.7
selenium>=4.15.0
webdriver-manager>=4.0.1
pandas>=2.1.0
numpy>=1.26.0
requests>=2.31.0
pytz>=2023.3
schedule>=1.2.1
